moq: pr notes
