+++ MOQ_DEV.md
@@ -1,4 +1,4 @@
-# MoQ local dev setup
+# MoQ local dev setup (TEMP - not for merge)
 
 End-to-end instructions for running the MoQ pipeline locally — a pipecat
 MoQ bot (acting as its own MOQ server) and the prebuilt React client
