+++ pyproject.toml
@@ -106,7 +106,7 @@ mistral = ["mistralai>=2.0.0,<3"]
 mlx-whisper = [ "mlx-whisper~=0.4.2" ]
 moondream = [ "accelerate~=1.10.0", "einops~=0.8.0", "pyvips[binary]~=3.0.0", "timm~=1.0.13", "transformers>=4.48.0,<6" ]
 moonshine = [ "moonshine-voice>=0.0.62" ]
-moq = [ "moq-rs>=0.2.13,<0.3", "cryptography>=43.0.0" ]
+moq = [ "moq-rs>=0.2.13,<1.0.0", "cryptography>=45.0.0" ]
 nebius = []
 neuphonic = []
 novita = []
