--- src/pipecat/transports/moq/__init__.py
@@ -1,30 +1,0 @@
-#
-# Copyright (c) 2024-2026, Daily
-#
-# SPDX-License-Identifier: BSD 2-Clause License
-#
-
-"""MOQ (Media over QUIC) transport implementation for Pipecat.
-
-This module provides MOQ transport functionality for real-time media streaming
-using the QUIC protocol, supporting low-latency audio and video transmission
-with pub/sub semantics.
-"""
-
-from pipecat.transports.moq.transport import (
-    MOQInputTransport,
-    MOQOutputTransport,
-    MOQParams,
-    MOQTrack,
-    MOQTrackType,
-    MOQTransport,
-)
-
-__all__ = [
-    "MOQInputTransport",
-    "MOQOutputTransport",
-    "MOQParams",
-    "MOQTrack",
-    "MOQTrackType",
-    "MOQTransport",
-]
